fix(bazar-list): handles non consistent api format
sometimes data.forms is an object of array, sometimes it's a object of objects. if the forms ids are sequential (0,1,2,3,4) it's interpreted as an standard array. But if you delete the form with id=2, then keys become (0,1,3,4) and then it's interpreted as a assoiative array -> i.e. converted into object