Same regex for all user name checks
The regex is rewriten to check if the user name respect the rules
(before the regex was checking for errors in the user name). This way
the same regex can be used for the HTML form and the PHP preg_match()
function.

Also update French and English translations to be more accurate about
the rules for the username valid format.